Copilot 리뷰 반영: CI 테스트 수정 및 createUserIfNotExists 안정성 강화
- parseStoredAuthUser를 authUserParser.ts로 분리하여 Firebase 의존성 없이 테스트 가능하게 수정
- createUserIfNotExists가 INITIAL_SESSION에서도 동작하도록 수정 (OAuth 리다이렉트 후 복원)
- 실패 시 userCreationAttempted 플래그를 리셋하여 재시도 가능하게 수정
- 로그아웃 시 플래그 리셋하여 다른 계정 로그인 시 사용자 생성 보장
