+++ src/shared/hooks/__tests__/parseStoredAuthUser.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { parseStoredAuthUser } from '../useAuth';
+import { parseStoredAuthUser } from '@/shared/utils/authUserParser';
 
 const VALID_UUID = '550e8400-e29b-41d4-a716-446655440000';
