+++ src/shared/hooks/useAuth.tsx
@@ -4,6 +4,7 @@ import { useContext, useState, useEffect, useRef, createContext } from 'react';
 import { getSupabaseClient } from '@/shared/api/supabaseClient';
 import { setSentryUser } from '@/sentry';
 import { mapToAuthUser } from '@/shared/auth/supabaseAuth';
+import { UUID_RE, parseStoredAuthUser } from '@/shared/utils/authUserParser';
 import { createUserIfNotExists } from '@/user/api/user';
 
 /**
@@ -32,21 +33,6 @@ export function useAuth() {
   return useContext(AuthContext);
 }
 
-export const UUID_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
-
-/** Parse stored AuthUser from localStorage with shape validation. */
-export function parseStoredAuthUser(raw: string | null): AuthUser | null {
-  if (!raw) return null;
-  try {
-    const parsed = JSON.parse(raw);
-    if (typeof parsed?.uid !== 'string' || !UUID_RE.test(parsed.uid)) {
-      return null;
-    }
-    return parsed as AuthUser;
-  } catch {
-    return null;
-  }
-}
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
   const [currentUser, setCurrentUser] = useState<AuthUser | null>(() => {
@@ -64,10 +50,17 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
     const { data: { subscription } } = supabase.auth.onAuthStateChange((event, session) => {
       const user = session?.user;
       const authUser = user && UUID_RE.test(user.id) ? mapToAuthUser(user) : null;
-      // Only create user on actual new logins, not on token refresh or tab re-focus
-      if (event === 'SIGNED_IN' && authUser && !userCreationAttempted.current) {
+      // Create user on actual sign-ins and initial session restore (OAuth redirect)
+      if ((event === 'SIGNED_IN' || event === 'INITIAL_SESSION') && authUser && !userCreationAttempted.current) {
         userCreationAttempted.current = true;
-        createUserIfNotExists(authUser).catch(console.error);
+        createUserIfNotExists(authUser).catch((error) => {
+          console.error(error);
+          userCreationAttempted.current = false;
+        });
+      }
+      // Reset flag on sign-out so a different user can trigger creation
+      if (!authUser) {
+        userCreationAttempted.current = false;
       }
       syncUserState(authUser, setCurrentUser);
       setLoading(false);
