+++ src/shared/utils/authUtils.ts
@@ -1,6 +1,6 @@
 import { getSupabaseClient } from '@/shared/api/supabaseClient';
 import type { AuthUser } from '@/shared/hooks/useAuth';
-import { UUID_RE } from '@/shared/hooks/useAuth';
+import { UUID_RE } from '@/shared/utils/authUserParser';
 import { mapToAuthUser } from '@/shared/auth/supabaseAuth';
 
 /**
